improve coverage

+++ core/src/main/java/org/apache/carbondata/core/datastore/page/encoding/adaptive/AdaptiveIntegralCodec.java
@@ -154,7 +154,7 @@ public class AdaptiveIntegralCodec extends AdaptiveCodec {
           encodedPage.putInt(rowId, (int) value);
           break;
         case LONG:
-          encodedPage.putLong(rowId, (long) value);
+          encodedPage.putLong(rowId, value);
           break;
         default:
           throw new RuntimeException("internal error: " + debugInfo());
